Paw_agent/requirements.txt
# FastAPI 및 웹 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 이미지 처리
Pillow==10.1.0
torch>=2.0.0
torchvision>=0.15.0

# OpenAI API
openai>=1.0.0

# 환경 변수 관리
python-dotenv==1.0.0

# 데이터 처리
numpy>=1.24.0

# Pydantic (FastAPI와 함께 설치되지만 명시)
pydantic>=2.0.0
